Move serializers to zookeeper-api (#1085)
Some serializer implementations (Chained, ByteArray) were never moved to zookeeper-api. This commit moves them out of helix-core and into zookeeper-api.

Note that this might cause older implementations of Helix applications to no longer build. Since this is an API-level change that simply moves the classes, the user can simply update the package path in the import statements and can make it build again.
